+++ lib/savon.rb
@@ -1,9 +1,10 @@
 module Savon
 
-  Error                = Class.new(RuntimeError)
-  InitializationError  = Class.new(Error)
-  UnknownOptionError   = Class.new(Error)
-  InvalidResponseError = Class.new(Error)
+  Error                 = Class.new(RuntimeError)
+  InitializationError   = Class.new(Error)
+  UnknownOptionError    = Class.new(Error)
+  UnknownOperationError = Class.new(Error)
+  InvalidResponseError  = Class.new(Error)
 
   def self.client(globals = {}, &block)
     Client.new(globals, &block)